set olcSecurity after olcLocalSSF
`olcSecurity` needs to come after `olcLocalSSF: 300`. Otherwise, the `simple_bind=128`
requirement causes a failure with an error message of "stronger confidentiality required".

Our conjecture is that this used to work because an older python and/or ansible
version iterated through the dict keys in a different order, so the `olcLocalSSF`
configuration change was applied before the `olcSecurity` change.